upworking

+++ upw/client.go
@@ -10,9 +10,9 @@ type AuthUser struct {
 
 // Client is the person advertising the job looking to hire the client
 type Client struct {
-	Country      string `json:"country"`
-	Feedback     int    `json:"feedback"`
-	ReviewsCount int    `json:"reviews_count"`
-	JobsPosted   int    `json:"jobs_posted"`
-	PastHires    int    `json:"past_hires"`
+	Country      string  `json:"country"`
+	Feedback     float64 `json:"feedback"`
+	ReviewsCount int     `json:"reviews_count"`
+	JobsPosted   int     `json:"jobs_posted"`
+	PastHires    int     `json:"past_hires"`
 }
